backend/requirements.txt
# FastAPI + GraphQL Backend Dependencies

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.214.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API integrations (reuse from CLI)
requests==2.31.0
certifi==2023.11.17

# Geospatial processing
geopy==2.4.0

# Data processing
pandas==2.1.4
numpy==1.24.4

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
structlog==23.2.0

# Async support
asyncpg==0.29.0  # For async PostgreSQL
httpx==0.25.2    # For async HTTP requests